Project of the NodeJS module is done

+++ module_NodeJS/project/project/handlers/requests_handlers.js
@@ -1,4 +1,6 @@
-const url = require("url"),
+const fs = require("fs"),
+      url = require("url"),
+      path = require("path"),
       queries = require("../models/queries"),
       dbPool = require("../models/db_connection").dbPool;
 
@@ -36,6 +38,10 @@ function displayContactPage(req, res, next) {
     res.render("contact", {title: "Contact"});
 }
 
+function displayLoginPage(req, res, next) {
+    res.render("login", {title: "Login"});
+}
+
 function checkLoginDetails(req, res, next) {
     if ((req.body.username === "admin") && (req.body.password === "12345")) {
         req.session.username = req.body.username;
@@ -62,6 +68,11 @@ function displayProjectPage(req, res, next) {
 }
 
 function displayProjectEditPage(req, res, next) {
+    if (req.session.username !== "admin") {
+        res.redirect("/login");
+        return;
+    }
+
     let projectId = req.params.id;
     queries.getProjectById(dbPool, projectId)
         .then((project) => {
@@ -79,6 +90,11 @@ function displayProjectEditPage(req, res, next) {
 }
 
 function deleteProject(req, res, next) {
+    if (req.session.username !== "admin") {
+        res.redirect("/login");
+        return;
+    }
+
     let projectId = req.params.id;
     queries.deleteProject(dbPool, projectId)
         .then(() => {
@@ -95,12 +111,16 @@ function displayCreateNewProjectPage(req, res, next) {
 }
 
 function addProject(req, res, next) {
+    let target_path = saveFile(req);
+
     let newProject = {
         title: req.body.title,
         author: req.body.author,
         category: req.body.category,
-        description: req.body.description
+        description: req.body.description,
+        src: target_path
     };
+
     queries.addProject(dbPool, newProject)
         .then(() => {
             res.redirect("/home");
@@ -112,12 +132,21 @@ function addProject(req, res, next) {
 }
 
 function updateProject(req, res, next) {
+    let target_path;
+
+    if (req.file) {
+        target_path = saveFile(req);
+    } else {
+        target_path = req.body.projectImage;
+    }
+
     let updatedProject = {
         id: req.params.id,
         title: req.body.title,
         author: req.body.author,
         category: req.body.category,
-        description: req.body.description
+        description: req.body.description,
+        src: target_path
     };
     queries.updateProject(dbPool, updatedProject)
         .then(() => {
@@ -136,10 +165,39 @@ function handleServerError(err, req, res, next) {
     }
 }
 
+// local (non exported) functions
+function saveFile(req) {
+    let tmp_path = req.file.path,
+        target_path = path.join(req.file.destination, req.file.originalname),
+        src = fs.createReadStream(tmp_path),
+        dest = fs.createWriteStream(target_path),
+        pathForDB = `/images/${req.file.originalname}`;
+
+    src.on("end", () => {
+        unlinkTmpPath();
+    });
+    src.on("error", (err) => {
+        unlinkTmpPath();
+    });
+
+    src.pipe(dest);
+
+    return pathForDB;
+
+    function unlinkTmpPath() {
+        fs.unlink(tmp_path, (err) => {
+            if (err) {
+                console.log(err);
+            }
+        });
+    }
+}
+
 module.exports = {
     goToHomePage,
     displayHomePage,
     displayContactPage,
+    displayLoginPage,
     checkLoginDetails,
     displayProjectPage,
     displayProjectEditPage,
